+++ TestProject/Player.h
@@ -5,7 +5,7 @@
 namespace nc {
 	class Player : public Character {
 	public:
-		Player(int _maxHP, int _mDefense, int _mAttack, int _pDefense, int _pAttack, int _evasion, int _position, std::string _name)
+		Player(int _maxHP, int _mDefense, int _mAttack, int _pDefense, int _pAttack, float _evasion, int _position, std::string _name)
 			: Character(_maxHP, _mDefense, _mAttack, _pDefense, _pAttack, _evasion, _position, _name) {
 		}
